+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add `bfd`, `dead_interval`, `hello_interval`, `mtu_ignore`, and `priority` attributes to `iosxe_interface_ospfv3` resource and data source
 - Add `iosxe_bgp_bmp_server` resource and data source
 - Add `ip_flow_monitors` and `ipv6_flow_monitors` attributes to `iosxe_interface_tunnel` resource and data source
 - Add `iosxe_device_tracking` resource and data source